backend/requirements.txt
# backend/requirements.txt
Django~=5.2.8             # Основний фреймворк
djangorestframework~=3.15 # Для створення API
psycopg2-binary~=2.9      # Адаптер для PostgreSQL
Pillow~=12.0              # Для роботи із зображеннями
djangorestframework-simplejwt~=5.3
djoser~=2.2
celery~=5.5.3                 # Для асинхронних завдань

